Don't initialize all events at the start of the simulation
Initializing -- running Geant4, that is -- for all events at the start of a
simulation will take up too much memory for configurations with many
events to simulate. Querying the thread pool's event queue for its size
, as is done in this commit, should keep memory usage relatively linear
throughout a simulation.

However, as a result, an additional thread is spawned: one thread
initializes the events, pushes them to the thread pool, and `threads_num`
workers are spawned to run them.

This additional master thread shouldn't cause performance issues; the
CPU will schedule it just fine.