objects: remove ROOT inheritance
This commit removes ROOT inheritance from the base Allpix² object. All
occurrences of TRef have been replaced with Object*. With this commit,
Allpix² performance is close to optimal for worker counts in the range
of 1 < n < 32.

Aside from leaking tons of memory, this commit likely breaks a lot of
stuff. Future work on an approach without ROOT inheritance would be:
    * Replace all new instances of Object* with std::shared_ptr<Object>;
      it likely scales much better.
    * Construct the final TObjects just before writing events to file
      instead.